requirements.txt
channels==3.0.4
channels-redis==4.2.0
daphne==3.0.2
Django==4.0.1
django-cors-headers==4.2.0
django-redis==5.4.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.9
redis==4.6.0
celery==5.3.6
PyMuPDF==1.23.26
docx2pdf==0.1.8
openpyxl==3.1.2
aspose-words==24.3.0
